Several upper bound estimates were extremely conservative, because they were upper quantiles of distributions which had at least one parameter which itself was randomly distributed, (compound distributions), and an upper quantile of this parameter distribution was being used.
The calculation of an upper bound on the number of delays ending up in the first delay bin has
been modified to account for the fact that the number of synapses is itself randomly distributed.

The calculation of an upper bound on the maximum delay has been modified to account for the fact that
the number of delays is randomly distributed.

Two functions have been introduced for the above, one to evaluate a cdf of a mixture distribution,
and another to evaluate the cdf of the maximum of a vector of random variables, where the length
of the vector is randomly distributed.

A function _row_synapses_distribution has been added to the connectors, returning a scipy distribution object.
The _estimate_max_row_synapses and _estimate_mean_row_synapses functions have been remove from the
connectors, since these can be calculated from the distribution returned by the above function.